// src/app/model/distribucion-asignacion.model.ts
export class DistribucionAsignacion {
    id: number;
    idUnidad: number;
    nomUnidad: string;
    idTambo: number;
    nomTambo: string;
    idTipocombustible: number;
    nomTipocombustible: string;
    idUnidadmedida: number;
    nomUnidadmedida: string;
    totalRequerimiento: number;
    precio: number;
    total: number;
    proveedor: string;

    nombreGestor: string;
    factura: string;
    direccionGrifo: string;

    ciudadAbastecimiento: string;
    distCiudadAbastecimiento: number;
    rucProveedor: string;
    observacion: string;
}